updated version of Tinasoft

+++ devel/mac/installer/install.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
-echo "Installing Tinasoft.."
+echo ""
+echo "**************************"
+echo " TINASOFT - MAC INSTALLER "
+echo "**************************"
+echo ""
+echo "  - installing Tinasoft to /Applications"
 base=/Applications/Tinasoft
 if [ -e $base ]
   then
-    echo "Uninstalling previous version.."
+    echo "  - uninstalling previous version.."
     rm -rf $base
 fi
 mkdir -p $base
 mv Tinasoft.app $base/
-echo "Moving config files out of the Tinasoft.app directory.."
+echo "  - installing config files to /Applications/Tinasoft/"
 mkdir -p $base/shared/
 mkdir -p $base/whitelists/
 mkdir -p $base/source_files/
@@ -16,12 +21,40 @@ mkdir -p $base/sessions/
 resources=$base/Tinasoft.app/Contents/Resources
 mv $resources/shared $base/
 mv $resources/source_files $base/
-mv $resources/whitelists $base/
+#mv $resources/whitelists $base/
 mv $resources/config_mac.yaml $base/config.yaml
-echo "Installing executable in Desktop and Applications.."
-chmod +x Tinasoft.sh
+echo "  - installing shortcut to /Applications/Tinasoft/"
 cp Tinasoft.sh $base/Tinasoft
+chmod +x $base/Tinasoft
+echo "  - installing shortcut to ~/Desktop/"
+mkdir -p ~/Desktop
 cp Tinasoft.sh ~/Desktop/Tinasoft
-echo "Cleaning.."
-rm *
-echo "you can now start the application by running 'Tinasoft'"
+chmod +x ~/Desktop/Tinasoft
+echo ""
+echo ""
+echo ""
+if [ -e /Desktop/Tinasoft ]
+  then
+    echo "*************************"
+    echo " INSTALLATION SUCCESSFUL "
+    echo "*************************"
+    echo ""
+    rm *
+    echo "Will now try to start server automatically. Please wait."
+    echo ""
+    echo "cd ~/Desktop"
+    cd ~/Desktop
+    echo "./Tinasoft"
+    ./Tinasoft
+  else
+    echo "!!!!!!!!!!!!!!!!!!!!!!!!!"
+    echo "!! INSTALLATION FAILED !!"
+    echo "!!!!!!!!!!!!!!!!!!!!!!!!!"
+    echo ""
+    echo "please send the installation history (copy/pasted from this terminal)"
+    echo "to julian.bilcke@gmail.com - Thanks a lot!"
+    if [ -e /Applications/Tinasoft ]
+      then
+        rm -Rf /Applications/Tinasoft
+    fi
+fi
